minor fixes and improvements
